fix: solves the Config filter problem
By changing configuration in the FilterScale atributte for CAN_FILTERSCALE_32BIT
the problem with the filter were solved. Also, this commits implements a
new functionality, a structure that tries to follow a "builder design
pattern" with the purpose of creating the mask and id for the filter
based in the standard can ids that the user wants to receive.

Signed-off-by: Carlos Veras <carlos.craveiro@usp.br>